refactor(hooks): move useAuth.js into src/hooks/
- Moved src/useAuth.js → src/hooks/useAuth.js
- Updated import in useAuth.js to use correct relative path (../authContext)
- Updated consumers: App.jsx, Dashboard.jsx, LoginPage.jsx

Hooks belong in src/hooks/ by React convention.
This makes the project structure easier to navigate.
