+++ fe-client/src/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useAuth} from "./useAuth.js";
+import {useAuth} from "./hooks/useAuth.js";
 import {useNavigate} from "react-router-dom";
 import UserProfile from "./components/UserProfile.jsx";
 import TripPlansList from "./components/TripPlansList.jsx";
