+++ fe-client/src/hooks/useAuth.js
@@ -1,5 +1,5 @@
 import {useContext} from "react";
-import {AuthContext} from "./authContext.jsx";
+import {AuthContext} from "../authContext.jsx";
 
 
 export const useAuth = () => {
